[+] Added Config Folder [+] Added SVG icons for items. [+] Added new widgets.
