+++ pubspec.yaml
@@ -31,8 +31,9 @@ dependencies:
   flutter:
     sdk: flutter
 
-  flutter_tree:
   flutter_fancy_tree_view:
+  flutter_svg:
+  draggable_resizer: ^0.0.2
   provider:
   #  vm_service: ^11.3.0
   # The following adds the Cupertino Icons font to your application.
@@ -74,7 +75,7 @@ flutter:
   # For details regarding adding assets from package dependencies, see
   # https://flutter.dev/assets-and-images/#from-packages
   assets:
-    - assets/images/
+    - assets/icons/
   # To add custom fonts to your application, add a fonts section here,
   # in this "flutter" section. Each entry in this list should have a
   # "family" key with the font family name, and a "fonts" key with a
